requirements.txt
absl-py==2.2.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.0
cython-bbox==0.1.5
dnspython==2.7.0
filelock==3.18.0
filterpy==1.4.5
fonttools==4.58.0
fsspec==2025.3.2
grpcio==1.71.0
h5py==3.13.0
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
loguru==0.7.3
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
motmetrics==1.4.0
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.0
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
pymongo==4.13.0
pyparsing==3.2.3
pyRAPL==0.2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
xmltodict==0.14.2
